Add initial .clang-format
Before my changes, { went to the end of the previous line, then in new code I
put it to the start of the next one. That's how I looked into some kind of
consistent style. But even before my changes, e.g.
lib/libclang-vim/location.hpp indented on namespace, but
lib/libclang-vim/tokenizer.hpp did not.

For the sake of simplicity, just stick to the LLVM style for now with
two exceptions as those were always consistent in this codebase.

For now, exclude test data from the changes.